Rewrite lvalue handling
Unfortunately the formerly designed lvalue architecture ended up
evaluating the indices multiple times when part of a substructure (once
on fetch, and again on store).

To correct this, add a third operation "change" which returns both the
current value and a function to change the underlying storage without
needing to re-evaluate the indices.

Also fix an oversight where we treated non-lvalue index and range
expressions as lvalues (to fetch their values) without supplying an
implementation.