Cabalize the distribution
Prepare the source for distribution by creating a .cabal file and
rewriting the Makefile to build the package with cabal.

Modify MOO.Version to return the version set in the .cabal file.

Move all the source code to a top-level directory 'src'.

Establish the license for the source code as BSD3.