Fix and revise exception handling
The existing implementation was flawed; an exception raised in a +d verb
could be passed up the call stack to a -d verb and then be turned into
an ordinary MOO value. Furthermore, an exception in a -d verb during the
evaluation of the parameters to for-loop and fork statements caused the
rest of the verb after the statement to be ignored.

To correct the first problem, we extend the Exception type to include a
copy of the debug bit from the verb frame where the exception occurred,
and refrain from collapsing exceptions into error values when set.

While we're extending the Exception type, we also incorporate the call
stack at the time of the exception, rather than passing this around
separately.