Fix problem with match() handling and transactions
It was unfortunately possible for an aborted STM transaction to wreak
havoc with our match() mutex, leading to eventual deadlock.

To correct this, we introduce helper C functions which are called via
FFI using "unsafe" semantics, so that the Haskell runtime blocks all
other running threads while each matching operation is performed.

It would be nice if PCRE did not rely on global state to make this
necessary, but alas.