Implement force_input()
Use our input queue machinery to add a line of input to either the
beginning or the end of the processing queue.

Note that because our queue is bounded, force_input() may raise E_QUOTA
if the queue is full when adding lines at the end of the queue. (This
doesn't happen when adding lines at the front of the queue.)