Refactor MOO binary string handling
Move all MOO binary string encoding and decoding to MOO.String in such a
way that (toBinary . fromBinary) does not suffer a round trip through
the encoding and decoding functions.

As part of this rewrite, also tidy up some loose ends with respect to
the eventual possibility of Unicode characters appearing within
strings. We allow only ASCII printing characters (including SPACE but
excluding HT) within MOO binary strings; all other characters will raise
E_INVARG if passed to decode_binary(). Similarly, only characters with
Unicode code points in the range 0-255 are allowed to be passed to
encode_binary(); this allows us to avoid the question of character
encoding for general Unicode. It is expected that additional built-in
functions will be added later to handle encodings such as UTF-8 and the
like.