Implement significant network and server features
Extend the networking infrastructure to implement most of the needed
features, and tie everything together into a usable server including
LambdaMOO-compatible option parsing of the command line.

Although we have introduced a "console" listener to the network stack,
it has problems and will probably need to be abandoned. Instead, it will
likely form the foundation for the Emergency Wizard Mode, as the MOO
command completion will work nicely there.