Use unsafeIOtoMOO a little more safely
It can happen that the IO action passed to unsafeIOtoMOO is executed
only once and then treated as a constant. To avoid this in our
implementation of the seconds_left() MOO built-in function, arrange for
the resulting value of the IO action to depend on a value from the
surrounding context.