Improve listDelete handling
Avoid O(n) thaw/create when unnecessary; tail/init alone can operate in
O(1) time.

Also, specialize the general case to minimize the number of elements to
move when the index to delete occurs in the first half of the list.

Note that unused element slots remain in the underlying vector storage,
so it would be good to implement a method for reclaiming the unused
memory at some point.