Extend listset() and listdelete()
Overload the listset() and listdelete() built-in functions to allow
setting and removing string-key/value pairs on association lists:

  alist = listset(alist, value, "key")  <=>  alist["key"] = value

There is no syntactic analog for listdelete(); this function is the
primary method for removing associations from a list:

  alist = {{"foo", 99}, {"bar", #30}, {"foo", 100}}

  listdelete(alist, "bar")  => {{"foo", 99}, {"foo", 100}}
  listdelete(alist, "Foo")  => {{"bar", #30}, {"foo", 100}}
  listdelete(alist, "baz")  => {{"foo", 99}, {"bar", #30}, {"foo", 100}}
  listdelete({2}, "foo")    (error) E_INVIND

Note that only the first instance of any duplicate key is removed.

N.B. The current implementation of listdelete() causes the resulting
list to rebuild its internal association hash table upon subsequent
access, which could be a performance bottleneck.