Refactor StrT representation
Try to reduce the MOO string footprint in the common case by isolating
the cached representations (as binary data or regular expression) into a
separate data record. This way we only need to store one thunk in the
string record for the function to create the cached representation(s) on
demand, instead of three thunks.

Also update the README to describe the special binary string handling
already implemented.