Fix match()/rmatch() multithreading
Our understanding of the "unsafe" semantics of Haskell's Foreign
Function Interface was incorrect. While it does suspend the execution of
other threads running on the same capability, it does not suspend any
threads that may be running on other capabilities. This makes it
unsuitable to use as a synchronization mechanism for modifying PCRE's
global pcre_callout variable.

Our solution is to ensure the same pcre_callout can be used for both
match() and rmatch(). Thus the synchronization issue becomes moot, and
we can instead use the FFI with "safe" semantics, allowing other threads
to continue running while the matching takes place.