Fix crypt() multithreading
As with the fix for match()/rmatch(), we cannot rely on the "unsafe"
semantics of Haskell's Foreign Function Interface to prevent reentrant
calls to the non-reentrant C crypt() function.

Instead, use an explicit mutex. This solution also allows us to call the
C crypt() directly, without the need for a helper function.

We had tried such a solution once before, but it failed in the STM
context because an aborted transaction could fail to restore the
bracketed mutex. Now that we have unsafeIOtoMOO to work around that
problem, this should no longer be an issue.