Fix "off/off of" preposition matching problem
There is a long-standing bug in LambdaMOO that prevents commands using
the "off of" preposition from being parsed correctly. The problem is
that "off" also matches such commands, and because this is tried first,
we never try matching with "off of". Consequently, the "of" ends up as
part of the indirect object string, where it interferes with object
matching.

Our solution is to reorder the prepositions as "off of/off" instead of
"off/off of", thus ensuring "off of" is tried first during command
matching.