Implement $server_options.protect_/prop/
If $server_options.protect_/prop/ exists and has a true value, then we
restrict both reading and writing the built-in property /prop/ (on any
object) to verbs running with wizard permissions, and raise E_PERM
otherwise.

Note that the LambdaMOO programmer's manual only specifies restricting
the reading of such protected properties, but it makes little sense not
to restrict writing also, and recent versions of the LambdaMOO server
code restrict writing as well as reading.