Enable SHA-3 hash functions
Now that SHA-3 is standardized, we can enable its use in the
value_hash(), string_hash(), and binary_hash() built-in functions. The
following algorithm names are now recognized when passed as the second
argument:

    "SHA3-224"
    "SHA3-256"
    "SHA3-384"
    "SHA3-512"

Unfortunately the implementations from the cryptohash package appear to
correspond with previous Keccak variants, and do not match the final
standard. Therefore we have switched to the cryptonite package to obtain
correct implementations.