Rewrite toText/toLiteral using Builder
This is potentially more efficient than the previous strategy of
concatenating Text values. We can also utilize the Builder
representation for the tostr() built-in function as well as for
unparsing literal expressions in verb code.

We add two new interface functions, toBuilder and toBuilder', which use
tostr() and toliteral() semantics, respectively.

The entire combination of toText/toBuilder/toBuilder'/toLiteral
functions are carefully constructed to adhere to the DRY principle, and
consequently appear to have some mutually recursive and cascading
relationships. However, the ordering of pattern matches and the case
coverage avoids any actual (unwanted) recursion or cascading beyond a
single level at runtime.

Thanks to Dylan Simon for the original suggestion.