Fix rmatch() problem with certain patterns
When a pattern starts with ".*", the rmatch() built-in function would
report the match as starting at the beginning of the string (assuming
the rest of the pattern matches). This is due to PCRE optimizations that
are appropriate for normal matching, but interfere with the intended
semantics of rmatch().

To fix the rmatch() behavior, we need to disable some of these
optimizations. As a current consequence, the use of JIT will be disabled
for all uses of rmatch(), though JIT can still be used for match().