Implement Emergency Wizard Mode
The mode is similar to that found in the LambdaMOO server, except that
arbitrary MOO commands are allowed, and the command line supports
editing, history, and completion.

Contrary to LambdaMOO, there is no abort command since database changes
are continuously committed. (One should keep a backup copy of the
database prior to entering Emergency Wizard Mode if there is a
possibility of making unwanted changes.)