Fix MOO.Parser crash on missing end* keyword
The code to illustrate error positions fails when the error occurs at
the end of the parsed text, such as for a missing endif/endfor/etc.
keyword.

To fix, add an empty line to the input text for illustration purposes. A
better fix will involve avoiding illustration of parse errors that occur
after the last input line.

Thanks to Konsta Ranta for reporting this issue.