Improve Emergency Wizard Mode
Rename the :quit command to :shutdown, to better reflect the resulting
action. Detect when the player has been disconnected (as from @quit) and
continue start-up in this case.

Rename the :wizard command to :wizard! to better reflect its hazardous
nature, and don't include the exclamation point in command completion so
that it must be typed explicitly.

Improve the prompt to reflect the current permissions in addition to the
current player, when these are different.

Document the Emergency Wizard Mode differences between EtaMOO and
LambdaMOO.